LayoutTests/ietestcenter/Javascript/15.4.4.14-9-b-i-6-expected.txt

15.4.4.14-9-b-i-6 (TestCases/chapter15/15.4/15.4.4/15.4.4.14/15.4.4.14-9-b-i-6.js) Array.prototype.indexOf when element to be retrieved is own data property that overrides an inherited accessor property on an Array-like object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
